framework region

+++ src/core/mamba_blocks.py
@@ -18,7 +18,7 @@ except ImportError:
 
 class BiBlock(nn.Module):
     def __init__(
-        self, dim, mixer_cls, norm_cls=nn.LayerNorm, fused_add_norm=False, residual_in_fp32=False
+        self, dim, mixer_cls, norm_cls = nn.LayerNorm, fused_add_norm = False, residual_in_fp32 = False
     ):
         """
         Simple block wrapping a mixer class with LayerNorm/RMSNorm and residual connection"
@@ -44,8 +44,9 @@ class BiBlock(nn.Module):
                 self.norm, (nn.LayerNorm, RMSNorm)
             ), "Only LayerNorm and RMSNorm are supported for fused_add_norm"
         self.LL_hidden = nn.Linear(288, 144)
+        
     def forward(
-        self, hidden_states: Tensor, residual: Optional[Tensor] = None, inference_params=None
+        self, hidden_states: Tensor, residual: Optional[Tensor] = None, inference_params = None
     ):
         r"""Pass the input through the encoder layer.
 
@@ -55,7 +56,7 @@ class BiBlock(nn.Module):
         """
         if not self.fused_add_norm:
             residual = (hidden_states + residual) if residual is not None else hidden_states
-            hidden_states = self.norm(residual.to(dtype=self.norm.weight.dtype))
+            hidden_states = self.norm(residual.to(dtype = self.norm.weight.dtype))
             if self.residual_in_fp32:
                 residual = residual.to(torch.float32)
         else:
@@ -64,57 +65,57 @@ class BiBlock(nn.Module):
                 hidden_states,
                 self.norm.weight,
                 self.norm.bias,
-                residual=residual,
-                prenorm=True,
-                residual_in_fp32=self.residual_in_fp32,
+                residual = residual,
+                prenorm = True,
+                residual_in_fp32 = self.residual_in_fp32,
                 eps=self.norm.eps,
             )
         f_hidden_states = hidden_states
         b_hidden_states = hidden_states.flip([1])
-        f_hidden_states = self.mixer(f_hidden_states, inference_params=inference_params)
-        b_hidden_states = self.inverse_mixer(b_hidden_states, inference_params=inference_params)
-        hidden_states = torch.cat((f_hidden_states, b_hidden_states), dim=-1)
+        f_hidden_states = self.mixer(f_hidden_states, inference_params = inference_params)
+        b_hidden_states = self.inverse_mixer(b_hidden_states, inference_params = inference_params)
+        hidden_states = torch.cat((f_hidden_states, b_hidden_states), dim = -1)
         hidden_states = self.LL_hidden(hidden_states)
         return hidden_states, residual
 
-    def allocate_inference_cache(self, batch_size, max_seqlen, dtype=None, **kwargs):
-        return self.mixer.allocate_inference_cache(batch_size, max_seqlen, dtype=dtype, **kwargs)
+    def allocate_inference_cache(self, batch_size, max_seqlen, dtype = None, **kwargs):
+        return self.mixer.allocate_inference_cache(batch_size, max_seqlen, dtype = dtype, **kwargs)
 
 
 def create_block(
     d_model,
-    ssm_cfg=None,
-    norm_epsilon=1e-5,
-    rms_norm=False,
-    residual_in_fp32=False,
-    fused_add_norm=False,
-    layer_idx=None,
-    device=None,
-    dtype=None,
-    bidirectional=False,
+    ssm_cfg = None,
+    norm_epsilon = 1e-5,
+    rms_norm = False,
+    residual_in_fp32 = False,
+    fused_add_norm = False,
+    layer_idx = None,
+    device = None,
+    dtype = None,
+    bidirectional = False,
 ):
     if ssm_cfg is None:
         ssm_cfg = {}
     factory_kwargs = {"device": device, "dtype": dtype}
-    mixer_cls = partial(Mamba, layer_idx=layer_idx, **ssm_cfg, **factory_kwargs)
+    mixer_cls = partial(Mamba, layer_idx = layer_idx, **ssm_cfg, **factory_kwargs)
     norm_cls = partial(
-        nn.LayerNorm if not rms_norm else RMSNorm, eps=norm_epsilon, **factory_kwargs
+        nn.LayerNorm if not rms_norm else RMSNorm, eps = norm_epsilon, **factory_kwargs
     )
     if not bidirectional:
         block = Block(
             d_model,
             mixer_cls,
-            norm_cls=norm_cls,
-            fused_add_norm=fused_add_norm,
-            residual_in_fp32=residual_in_fp32,
+            norm_cls = norm_cls,
+            fused_add_norm = fused_add_norm,
+            residual_in_fp32 = residual_in_fp32,
         )
     else:
         block = BiBlock(
             d_model,
             mixer_cls,
-            norm_cls=norm_cls,
-            fused_add_norm=fused_add_norm,
-            residual_in_fp32=residual_in_fp32,
+            norm_cls = norm_cls,
+            fused_add_norm = fused_add_norm,
+            residual_in_fp32 = residual_in_fp32,
         )
     block.layer_idx = layer_idx
     return block
@@ -124,16 +125,16 @@ def create_block(
 def _init_weights(
     module,
     n_layer,
-    initializer_range=0.02,  # Now only used for embedding layer.
-    rescale_prenorm_residual=True,
-    n_residuals_per_layer=1,  # Change to 2 if we have MLP
+    initializer_range = 0.02,  # Now only used for embedding layer.
+    rescale_prenorm_residual = True,
+    n_residuals_per_layer = 1,  # Change to 2 if we have MLP
 ):
     if isinstance(module, nn.Linear):
         if module.bias is not None:
             if not getattr(module.bias, "_no_reinit", False):
                 nn.init.zeros_(module.bias)
     elif isinstance(module, nn.Embedding):
-        nn.init.normal_(module.weight, std=initializer_range)
+        nn.init.normal_(module.weight, std = initializer_range)
 
     if rescale_prenorm_residual:
         # Reinitialize selected weights subject to the OpenAI GPT-2 Paper Scheme:
@@ -148,7 +149,7 @@ def _init_weights(
                 # Following Pytorch init, except scale by 1/sqrt(2 * n_layer)
                 # We need to reinit p since this code could be called multiple times
                 # Having just p *= scale would repeatedly scale it down
-                nn.init.kaiming_uniform_(p, a=math.sqrt(5))
+                nn.init.kaiming_uniform_(p, a = math.sqrt(5))
                 with torch.no_grad():
                     p /= math.sqrt(n_residuals_per_layer * n_layer)
 
@@ -159,15 +160,15 @@ class BiMixerModel(nn.Module):
         d_model: int,
         n_layer: int,
         # vocab_size: int,
-        ssm_cfg=None,
+        ssm_cfg = None,
         norm_epsilon: float = 1e-5,
         rms_norm: bool = False,
-        if_bidirectional=True,
-        initializer_cfg=None,
-        fused_add_norm=False,
-        residual_in_fp32=False,
-        device=None,
-        dtype=None,
+        if_bidirectional = True,
+        initializer_cfg = None,
+        fused_add_norm = False,
+        residual_in_fp32 = False,
+        device = None,
+        dtype = None,
     ) -> None:
         factory_kwargs = {"device": device, "dtype": dtype}
         super().__init__()
@@ -188,66 +189,67 @@ class BiMixerModel(nn.Module):
             [
                 create_block(
                     d_model,
-                    ssm_cfg=ssm_cfg,
-                    norm_epsilon=norm_epsilon,
-                    rms_norm=rms_norm,
-                    residual_in_fp32=residual_in_fp32,
-                    fused_add_norm=fused_add_norm,
-                    layer_idx=i,
-                    bidirectional=False,
+                    ssm_cfg = ssm_cfg,
+                    norm_epsilon = norm_epsilon,
+                    rms_norm = rms_norm,
+                    residual_in_fp32 = residual_in_fp32,
+                    fused_add_norm = fused_add_norm,
+                    layer_idx = i,
+                    bidirectional = False,
                     **factory_kwargs,
                 )
                 for i in range(n_layer)
             ]
         )
         
-
         self.norm_f = (nn.LayerNorm if not rms_norm else RMSNorm)(
-            d_model, eps=norm_epsilon, **factory_kwargs
+            d_model, eps = norm_epsilon, **factory_kwargs
         )
 
         self.apply(
             partial(
                 _init_weights,
-                n_layer=n_layer,
+                n_layer = n_layer,
                 **(initializer_cfg if initializer_cfg is not None else {}),
             )
         )
         self.f_attention_pool = nn.Linear(d_model, 1)
-        self.dropout = nn.Dropout(p=0.1)
+        self.dropout = nn.Dropout(p = 0.1)
         self.classifier = nn.Linear(d_model, 2)
-    def allocate_inference_cache(self, batch_size, max_seqlen, dtype=None, **kwargs):
+
+    def allocate_inference_cache(self, batch_size, max_seqlen, dtype = None, **kwargs):
         return {
-            i: layer.allocate_inference_cache(batch_size, max_seqlen, dtype=dtype, **kwargs)
+            i: layer.allocate_inference_cache(batch_size, max_seqlen, dtype = dtype, **kwargs)
             for i, layer in enumerate(self.layers)
         }
 
-    def forward(self, x, inference_params=None):
+    def forward(self, x, inference_params = None):
         hidden_states = self.dropout(x)
         residual = None
         for layer in self.layers:
             hidden_states, residual = layer(
-                hidden_states, residual, inference_params=inference_params
+                hidden_states, residual, inference_params = inference_params
             )
         if not self.fused_add_norm:
             residual = (hidden_states + residual) if residual is not None else hidden_states
-            hidden_states = self.norm_f(residual.to(dtype=self.norm_f.weight.dtype))
+            hidden_states = self.norm_f(residual.to(dtype = self.norm_f.weight.dtype))
         else:
             fused_add_norm_fn = rms_norm_fn if isinstance(self.norm_f, RMSNorm) else layer_norm_fn
             hidden_states = fused_add_norm_fn(
                 hidden_states,
                 self.norm_f.weight,
                 self.norm_f.bias,
-                eps=self.norm_f.eps,
-                residual=residual,
-                prenorm=False,
-                residual_in_fp32=self.residual_in_fp32,
+                eps = self.norm_f.eps,
+                residual = residual,
+                prenorm = False,
+                residual_in_fp32 = self.residual_in_fp32,
             )
 
         hidden_states = torch.matmul(F.softmax(self.f_attention_pool(
-            hidden_states), dim=1).transpose(-1, -2), hidden_states).squeeze(-2)
+            hidden_states), dim = 1).transpose(-1, -2), hidden_states).squeeze(-2)
         hidden_states = self.dropout(hidden_states)
         hidden_states = self.classifier(hidden_states)
+
         return hidden_states
    
     
@@ -257,15 +259,15 @@ class MixerModel(nn.Module):
         d_model: int,
         n_layer: int,
         # vocab_size: int,
-        ssm_cfg=None,
+        ssm_cfg = None,
         norm_epsilon: float = 1e-5,
         rms_norm: bool = False,
-        if_bidirectional=True,
-        initializer_cfg=None,
-        fused_add_norm=False,
-        residual_in_fp32=False,
-        device=None,
-        dtype=None,
+        if_bidirectional = True,
+        initializer_cfg = None,
+        fused_add_norm = False,
+        residual_in_fp32 = False,
+        device = None,
+        dtype = None,
     ) -> None:
         factory_kwargs = {"device": device, "dtype": dtype}
         super().__init__()
@@ -286,27 +288,28 @@ class MixerModel(nn.Module):
             [
                 create_block(
                     d_model,
-                    ssm_cfg=ssm_cfg,
-                    norm_epsilon=norm_epsilon,
-                    rms_norm=rms_norm,
-                    residual_in_fp32=residual_in_fp32,
-                    fused_add_norm=fused_add_norm,
-                    layer_idx=i,
+                    ssm_cfg = ssm_cfg,
+                    norm_epsilon = norm_epsilon,
+                    rms_norm = rms_norm,
+                    residual_in_fp32 = residual_in_fp32,
+                    fused_add_norm = fused_add_norm,
+                    layer_idx = i,
                     **factory_kwargs,
                 )
                 for i in range(n_layer)
             ]
         )
+
         self.backward_layers = nn.ModuleList(
             [
                 create_block(
                     d_model,
-                    ssm_cfg=ssm_cfg,
-                    norm_epsilon=norm_epsilon,
-                    rms_norm=rms_norm,
-                    residual_in_fp32=residual_in_fp32,
-                    fused_add_norm=fused_add_norm,
-                    layer_idx=i,
+                    ssm_cfg = ssm_cfg,
+                    norm_epsilon = norm_epsilon,
+                    rms_norm = rms_norm,
+                    residual_in_fp32 = residual_in_fp32,
+                    fused_add_norm = fused_add_norm,
+                    layer_idx = i,
                     **factory_kwargs,
                 )
                 for i in range(n_layer)
@@ -314,49 +317,50 @@ class MixerModel(nn.Module):
         )
 
         self.norm_f = (nn.LayerNorm if not rms_norm else RMSNorm)(
-            d_model, eps=norm_epsilon, **factory_kwargs
+            d_model, eps = norm_epsilon, **factory_kwargs
         )
 
         self.apply(
             partial(
                 _init_weights,
-                n_layer=n_layer,
+                n_layer = n_layer,
                 **(initializer_cfg if initializer_cfg is not None else {}),
             )
         )
         self.f_attention_pool = nn.Linear(d_model, 1)
         self.b_attention_pool = nn.Linear(d_model, 1)
         self.LL = nn.Linear(d_model * 2, d_model)
-        self.dropout = nn.Dropout(p=0.1)
+        self.dropout = nn.Dropout(p = 0.1)
         self.classifier = nn.Linear(d_model, 2)
-    def allocate_inference_cache(self, batch_size, max_seqlen, dtype=None, **kwargs):
+
+    def allocate_inference_cache(self, batch_size, max_seqlen, dtype = None, **kwargs):
         return {
-            i: layer.allocate_inference_cache(batch_size, max_seqlen, dtype=dtype, **kwargs)
+            i: layer.allocate_inference_cache(batch_size, max_seqlen, dtype = dtype, **kwargs)
             for i, layer in enumerate(self.layers)
         }
 
-    def forward(self, x, inference_params=None):
+    def forward(self, x, inference_params = None):
         hidden_states = self.dropout(x)
         
         if not self.if_bidirectional:
             residual = None
             for layer in self.layers:
                 hidden_states, residual = layer(
-                    hidden_states, residual, inference_params=inference_params
+                    hidden_states, residual, inference_params = inference_params
                 )
             if not self.fused_add_norm:
                 residual = (hidden_states + residual) if residual is not None else hidden_states
-                hidden_states = self.norm_f(residual.to(dtype=self.norm_f.weight.dtype))
+                hidden_states = self.norm_f(residual.to(dtype = self.norm_f.weight.dtype))
             else:
                 fused_add_norm_fn = rms_norm_fn if isinstance(self.norm_f, RMSNorm) else layer_norm_fn
                 hidden_states = fused_add_norm_fn(
                     hidden_states,
                     self.norm_f.weight,
                     self.norm_f.bias,
-                    eps=self.norm_f.eps,
-                    residual=residual,
-                    prenorm=False,
-                    residual_in_fp32=self.residual_in_fp32,
+                    eps = self.norm_f.eps,
+                    residual = residual,
+                    prenorm = False,
+                    residual_in_fp32 = self.residual_in_fp32,
                 )
         else:
             f_hidden_states = hidden_states
@@ -372,7 +376,7 @@ class MixerModel(nn.Module):
                 )
             if not self.fused_add_norm:
                 f_residual = (f_hidden_states + f_residual) if f_residual is not None else f_hidden_states
-                f_hidden_states = self.norm_f(f_residual.to(dtype=self.norm_f.weight.dtype))
+                f_hidden_states = self.norm_f(f_residual.to(dtype = self.norm_f.weight.dtype))
             else:
                 fused_add_norm_fn = rms_norm_fn if isinstance(self.norm_f, RMSNorm) else layer_norm_fn
                 # Use long-range features as the key for residual connections
@@ -380,14 +384,14 @@ class MixerModel(nn.Module):
                     hidden_states,
                     self.norm_f.weight,
                     self.norm_f.bias,
-                    eps=self.norm_f.eps,
-                    residual=f_residual,
-                    prenorm=False,
-                    residual_in_fp32=self.residual_in_fp32,
+                    eps = self.norm_f.eps,
+                    residual = f_residual,
+                    prenorm = False,
+                    residual_in_fp32 = self.residual_in_fp32,
                 )
             if not self.fused_add_norm:
                 b_residual = (b_hidden_states + b_residual) if b_residual is not None else b_hidden_states
-                b_hidden_states = self.norm_f(b_residual.to(dtype=self.norm_f.weight.dtype))
+                b_hidden_states = self.norm_f(b_residual.to(dtype = self.norm_f.weight.dtype))
             else:
                 fused_add_norm_fn = rms_norm_fn if isinstance(self.norm_f, RMSNorm) else layer_norm_fn
                 # Use long-range features as the key for residual connections
@@ -395,17 +399,17 @@ class MixerModel(nn.Module):
                     hidden_states,
                     self.norm_f.weight,
                     self.norm_f.bias,
-                    eps=self.norm_f.eps,
-                    residual=b_residual,
-                    prenorm=False,
-                    residual_in_fp32=self.residual_in_fp32,
+                    eps = self.norm_f.eps,
+                    residual = b_residual,
+                    prenorm = False,
+                    residual_in_fp32 = self.residual_in_fp32,
                 )
         f_hidden_states = torch.matmul(F.softmax(self.f_attention_pool(
-            f_hidden_states), dim=1).transpose(-1, -2), f_hidden_states).squeeze(-2)
+            f_hidden_states), dim = 1).transpose(-1, -2), f_hidden_states).squeeze(-2)
         b_hidden_states = torch.matmul(F.softmax(self.b_attention_pool(
-            b_hidden_states), dim=1).transpose(-1, -2), b_hidden_states).squeeze(-2)
-        hidden_states = torch.cat((f_hidden_states, b_hidden_states), dim=1)
-        hidden_states = self.LL(hidden_states)
+            b_hidden_states), dim = 1).transpose(-1, -2), b_hidden_states).squeeze(-2)
+        hidden_states = torch.cat((f_hidden_states, b_hidden_states), dim = 1)
+        hidden_states = self.LL(hidden_states) 
         hidden_states = self.dropout(hidden_states)
         hidden_states = self.classifier(hidden_states)
